review: self-review fixes - honest docs, lossless version, render wording
From a pre-review pass over the whole diff:

- propose_against_pg / with_trace / PgProposalOutcome rustdoc now
  names the recording side-effect: embedders treating rejection as
  read-rollback learn about the post-rollback insert from the public
  contract, not a private helper's doc.
- The require arm renders its expression once again (the bind arm's
  shape); the stale render-once comment is true again.
- invariant_version becomes bigint/i64 end to end - the column has
  shipped nowhere yet, so the lossless conversion is free now and
  never again.
- The kind vocabulary is shared consts so writer and reader cannot
  drift; the CHECK constraint pins the same three values.
- A transformation that was only ever refused renders as 'never
  committed a transition' - '0 transition(s)' beside a refusal count
  read as never-proposed, the opposite of true.
- inspect coverage's CLI help mentions the constrained verdict;
  adapter.py's coverage docstring likewise (package regenerated).
- Migration 005 names its failure mode: unapplied, the breakage
  lands on the refusal path embedders exercise constantly.
- The tutorial's 'nothing was written' becomes 'nothing was
  admitted' plus the rejection-log aside - the one stale-prose hit.
- design-history stops overclaiming: the batch path is now pinned
  end-to-end (the batch CLI test checks the log), compensation is
  named as structural-not-tested, and the courtesy line's
  display-text parse is acknowledged as the deliberate residual.

+++ crates/morpholog-core/sql/migrations/005_rejections.sql
@@ -6,6 +6,13 @@
 --
 -- Idempotent: re-running is safe.
 --
+-- Until applied, every REJECTION (the normal business refusal path)
+-- and every `inspect coverage` / `inspect rejections` surfaces as an
+-- operational error, because the recording insert and the replay
+-- read hit a missing table. Unlike earlier migrations, the breakage
+-- lands on the refusal path embedders exercise constantly - apply
+-- this before upgrading the binary.
+--
 -- What this migration does:
 --   Creates morpholog.rejections, one row per refused proposal,
 --   recording who proposed what and which rule refused it (kind +
@@ -28,7 +35,7 @@ CREATE TABLE IF NOT EXISTS rejections (
     actor                jsonb        NOT NULL,
     kind                 text         NOT NULL CHECK (kind IN ('invariant', 'require', 'bind')),
     rule                 text         NOT NULL,
-    invariant_version    int,
+    invariant_version    bigint,
     reason               text         NOT NULL,
     rejected_at          timestamptz  NOT NULL DEFAULT now()
 );
